requirements.txt
langgraph
langchain
langchain-anthropic
langchain-core
langchain-openai
google-auth-oauthlib
google-api-python-client
python-dotenv
jupyter
pytest